#998 test: extended coverage

+++ plugins/BEdita/Core/tests/TestCase/Shell/BeditaShellTest.php
@@ -3,6 +3,8 @@ namespace BEdita\Core\Test\TestCase\Shell;
 
 use BEdita\Core\Shell\BeditaShell;
 use BEdita\Core\Utility\Database;
+use Cake\Core\Configure;
+use Cake\Datasource\ConnectionManager;
 use Cake\ORM\TableRegistry;
 use Cake\TestSuite\TestCase;
 
@@ -29,6 +31,26 @@ class BeditaShellTest extends TestCase
     public $BeditaShell;
 
     /**
+     * Fixtures
+     *
+     * @var array
+     */
+    public $fixtures = [
+        'plugin.BEdita/Core.config',
+        'plugin.BEdita/Core.users',
+    ];
+
+    /**
+     * @inheritDoc
+     */
+    public $autoFixtures = false;
+
+    /**
+     * Exclude from drop tables action
+     */
+    public $excludeFromDrop = [];
+
+    /**
      * setUp method
      *
      * @return void
@@ -41,6 +63,24 @@ class BeditaShellTest extends TestCase
         $this->BeditaShell->initialize();
     }
 
+    protected function fakeDbSetup($configName, $prefix = '__BE4_', $suffix = '__')
+    {
+        $fake = [
+            'className' => 'Cake\Database\Connection',
+            'host' => $prefix . 'DB_HOST' . $suffix,
+            'username' => $prefix . 'DB_USERNAME' . $suffix,
+            'password' => $prefix . 'DB_PASSWORD' . $suffix,
+            'database' => $prefix . 'DB_DATABASE' . $suffix,
+        ];
+        $info = Database::basicInfo();
+        $fake = array_merge($info, $fake);
+        Configure::write('Datasources.' . $configName, $fake);
+        ConnectionManager::config($configName, $fake);
+
+        return $fake;
+    }
+
+
     /**
      * tearDown method
      *
@@ -48,6 +88,7 @@ class BeditaShellTest extends TestCase
      */
     public function tearDown()
     {
+        $this->dbCleanup();
         unset($this->BeditaShell);
 
         parent::tearDown();
@@ -71,7 +112,7 @@ class BeditaShellTest extends TestCase
     {
         return [
             'noSetup' => [
-                ['y', 'n', 'n'],
+                ['y', 'n', 'y'],
                 [],
                 ['pippo', 'pippo']
             ],
@@ -124,7 +165,7 @@ class BeditaShellTest extends TestCase
             $this->markTestSkipped('MySQL only supported (for now)');
         }
 
-        $this->BeditaShell->setup();
+        $res = $this->BeditaShell->setup();
 
         if ($userPass[0]) {
             $usersTable = TableRegistry::get('Users');
@@ -133,11 +174,197 @@ class BeditaShellTest extends TestCase
             $this->assertEquals($userPass[0], $user->username);
         }
 
+        $this->dbCleanup();
+    }
+
+    /**
+     * Test setup with fake connection data
+     *
+     * @return void
+     */
+    public function testFake()
+    {
+        $this->fixtureManager->shutDown();
+
+        $this->fakeDbSetup('__test-temp__');
+        ConnectionManager::alias('__test-temp__', 'default');
+
+        $res = $this->BeditaShell->setup();
+        $this->assertFalse($res);
+
+        ConnectionManager::alias('test', 'default');
+        ConnectionManager::drop('__test-temp__');
+        $this->dbCleanup();
+    }
+
+    /**
+     * Test setup with fake data
+     *
+     * @return void
+     */
+    public function testFake2()
+    {
+        $this->fixtureManager->shutDown();
+
+        $fakeParams = $this->fakeDbSetup('__test-temp__');
+        ConnectionManager::alias('__test-temp__', 'default');
+
+        $map = [
+            ['Host?', null, $fakeParams['host']],
+            ['Database?', null, $fakeParams['database']],
+            ['Username?', null, $fakeParams['username']],
+            ['Password?', null, $fakeParams['password']],
+        ];
+        $this->io->method('ask')
+             ->will($this->returnValueMap($map));
+
+        $res = $this->BeditaShell->setup();
+        $this->assertFalse($res);
+
+        $mapChoice = [
+            ['Proceed with setup?', ['y', 'n'], 'n', 'y'],
+        ];
+        $this->io->method('askChoice')
+             ->will($this->returnValueMap($mapChoice));
+
+        $res = $this->BeditaShell->setup();
+        $this->assertFalse($res);
+
+
+        ConnectionManager::alias('test', 'default');
+        ConnectionManager::drop('__test-temp__');
+        $this->dbCleanup();
+    }
+
+
+    /**
+     * Test save connection to file failure
+     *
+     * @return void
+     */
+    public function testFake3()
+    {
+        $this->fixtureManager->shutDown();
+
+        $info = Database::basicInfo();
+
+        $fakeParams = $this->fakeDbSetup('__test-temp__');
+        ConnectionManager::alias('__test-temp__', 'default');
+
+        $mapChoice = [
+            ['Proceed with setup?', ['y', 'n'], 'n', 'y'],
+            ['Proceed with database schema and data initialization?', ['y', 'n'], 'n', 'y'],
+        ];
+        $this->io->method('askChoice')
+             ->will($this->returnValueMap($mapChoice));
+
+        $map = [
+            ['Host?', null, $info['host']],
+            ['Database?', null, $info['database']],
+            ['Username?', null, $info['username']],
+            ['Password?', null, $info['password']],
+            ['username: ', null, 'username'],
+            ['password: ', null, '42']
+        ];
+        $this->io->method('ask')
+             ->will($this->returnValueMap($map));
+
+        $fakeCfg = TMP . '_test_app.php.tmp';
+
+        $res = copy(CONFIG . 'app.default.php', $fakeCfg);
+        $this->assertTrue($res);
+        $this->BeditaShell->configPath = $fakeCfg;
+
+        $res = $this->BeditaShell->setup();
+        $this->assertTrue($res);
+
+        $res = unlink($fakeCfg);
+        $this->assertTrue($res);
+        $res = $this->BeditaShell->setup();
+        $this->assertFalse($res);
+
+        $res = touch($fakeCfg);
+        $this->assertTrue($res);
+        $res = $this->BeditaShell->setup();
+        $this->assertFalse($res);
+
+        $this->dbCleanup();
+
+        ConnectionManager::alias('test', 'default');
+        ConnectionManager::drop('__test-temp__');
+    }
+
+
+    /**
+     * Test init schema failure
+     *
+     * @return void
+     */
+    public function testInitSchemaFailure()
+    {
+        $this->fixtureManager->shutDown();
+        $this->loadFixtures('Config');
+
+        $info = Database::basicInfo();
+
+        $mapChoice = [
+            ['Proceed with setup?', ['y', 'n'], 'n', 'y'],
+        ];
+        $this->io->method('askChoice')
+             ->will($this->returnValueMap($mapChoice));
+
+        $res = $this->BeditaShell->setup();
+        $this->assertFalse($res);
+
+        $this->excludeFromDrop[] = 'config';
+        $this->dbCleanup();
+    }
+
+    /**
+     * Test admin user setup failure
+     *
+     * @return void
+     */
+    public function testAdminUserFailure()
+    {
+        $this->fixtureManager->shutDown();
+
+        $mapChoice = [
+            ['Proceed with setup?', ['y', 'n'], 'n', 'y'],
+            ['Proceed with database schema and data initialization?', ['y', 'n'], 'n', 'y'],
+        ];
+        $this->io->method('askChoice')
+             ->will($this->returnValueMap($mapChoice));
+
+        $info = Database::basicInfo();
+        $map = [
+            ['username: ', null, ''],
+            ['password: ', null, '']
+        ];
+        $this->io->method('ask')
+             ->will($this->returnValueMap($map));
+
+        $this->BeditaShell->defaultUsername = 'somename';
+        $res = $this->BeditaShell->setup();
+        $this->assertTrue($res);
+
+        $this->BeditaShell->defaultUsername = 'bedita';
+        $this->expectException('\Cake\Console\Exception\StopException');
+        $this->BeditaShell->setup();
+
+    }
+
+
+    protected function dbCleanup()
+    {
         $schema = Database::currentSchema();
         if (!empty($schema)) {
             $res = Database::executeTransaction($this->dropTablesSql($schema));
             $this->assertNotEmpty($res);
             $this->assertEquals($res['success'], true);
+        }
+        if (ConnectionManager::config('__temp_setup__')) {
+            ConnectionManager::drop('__temp_setup__');
         }
     }
 
@@ -151,7 +378,9 @@ class BeditaShellTest extends TestCase
     {
         $sql[] = 'SET FOREIGN_KEY_CHECKS=0;';
         foreach ($schema as $k => $v) {
-            $sql[] = 'DROP TABLE IF EXISTS ' . $k;
+            if (!in_array($k, $this->excludeFromDrop)) {
+                $sql[] = 'DROP TABLE IF EXISTS ' . $k;
+            }
         }
         $sql[] = 'SET FOREIGN_KEY_CHECKS=1;';
